Build syncclient with dependency injection system
